Completion value of a finally block should not be ignored if completion is abrupt https://bugs.webkit.org/show_bug.cgi?id=221238
Reviewed by Yusuke Suzuki.

JSTests:

* stress/completion-value.js:
Correct tests to match spec.

* test262/expectations.yaml:
Mark two test cases as passing.

Source/JavaScriptCore:

Per https://tc39.es/ecma262/#sec-try-statement-runtime-semantics-evaluation,
the completion value of a finally block is ignored *just* when it is a normal completion,
but we've been ignoring it in all cases.

This patch ensures that when a finally block is exited with a break or continue statement,
its completion value is used as the completion value for the entire TryStatement.

(Note: This behavior will be important for the upcoming "do expressions" proposal,
as it is effectively a reification of completion values.)

* bytecompiler/NodesCodegen.cpp:
(JSC::TryNode::emitBytecode):
1. Don't throw away the result of evaluating the finally block.
2. Only use the try or catch block result if we make it all the way to the end of the finally block.


Canonical link: https://commits.webkit.org/233619@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@272243 268f45cc-cd09-0410-ab3c-d52691b4dbfc