http/wpt/mediarecorder/mute-tracks.html is flaky https://bugs.webkit.org/show_bug.cgi?id=221195
Reviewed by Eric Carlson.

In case bot is slow, we were doing more than one analysis.
For each analysis, we were creating a new MediaElementSourceNode for the same video element which is not allowed.
Instead, do the AudioContext setup once and run the analysis steps as long as needed.

* http/wpt/mediarecorder/mute-tracks-expected.txt:
* http/wpt/mediarecorder/mute-tracks.html:


Canonical link: https://commits.webkit.org/233627@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@272296 268f45cc-cd09-0410-ab3c-d52691b4dbfc