Audio is sometimes distorted when using createMediaElementSource https://bugs.webkit.org/show_bug.cgi?id=221553
Reviewed by Eric Carlson.

AudioSourceProviderAVFObjC::provideInput() was calculating the number of
frames available in the RingBuffer like so:
`endFrame - (m_readCount + writeAheadCount)`

The issue is that when writeAheadCount is greater than 0, (m_readCount + writeAheadCount)
can become greater than endFrame, causing us to end up with a super large number of
available frames. This can lead us to render garbage because we may be reading more
frames than available the RingBuffer.

I have also made a change in the case where rendering has already started were we ignore
writeAheadCount and render whatever we have in the RingBuffer. After we've started rendering,
deciding not to render because of writeAheadCount, even though there are available frames
in the RingBuffer, is really hurtful to user experience. This was reproducible on
https://www.waveplayer.info/createmediaelementsource-test/ where we would start rendering
for a bit, then output silence and then pick up rendering again (this time for good).

* platform/graphics/avfoundation/AudioSourceProviderAVFObjC.mm:
(WebCore::AudioSourceProviderAVFObjC::provideInput):


Canonical link: https://commits.webkit.org/233970@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@272730 268f45cc-cd09-0410-ab3c-d52691b4dbfc