[CMake] WEBKIT_EXECUTABLE can incorrectly link framework https://bugs.webkit.org/show_bug.cgi?id=221703
Reviewed by Michael Catanzaro.

.:

After r272484 which added an additional test for TestJavaScriptCore the PlayStation build
began failing with an unresolved symbol in bmalloc. On PlayStation both WTF and bmalloc
are linked into JavaScriptCore. That library was building successfully which implied that
WTF was being erroniously linked into TestJavaScriptCore.

Inside _WEBKIT_EXECUTABLE_LINK_FRAMEWORK there was an invalid conditional which was causing
WTF to be linked. Fixed the logic to ensure that if the requested framework is linked into
another framework that is being linked then it is not added to the linker.

* Source/cmake/WebKitMacros.cmake:

Source/JavaScriptCore:

GTK compiles JavaScriptCore as a SHARED library but the symbols from WTF are not being
properly exposed from it. Workaround this by linking WTF directly until GTK turns on
hidden visibility and properly exports symbols.

* shell/PlatformGTK.cmake:

Tools:

GTK compiles JavaScriptCore as a SHARED library but the symbols from WTF are not being
properly exposed from it. Workaround this by linking WTF directly until GTK turns on
hidden visibility and properly exports symbols.

* TestWebKitAPI/CMakeLists.txt:
* TestWebKitAPI/PlatformGTK.cmake:


Canonical link: https://commits.webkit.org/233972@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@272732 268f45cc-cd09-0410-ab3c-d52691b4dbfc