[Cocoa] IPC decoder is using decoded size to allocate memory for an array https://bugs.webkit.org/show_bug.cgi?id=221773
Reviewed by Geoffrey Garen.

* Shared/Cocoa/ArgumentCodersCocoa.mm:
(IPC::decodeArrayInternal): As with other similar structures, such a Vector and
CFArray, don't use the size to preallocate space when decoding an NSArray. The
decoded size is potentially incorrect, which we will discover indirectly when
decoding the array elements; we can't safely use the size to make a choice about
allocating memory beforehand.

Canonical link: https://commits.webkit.org/233984@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@272744 268f45cc-cd09-0410-ab3c-d52691b4dbfc