[LFC][Integration] Make accessibility/aria-modal-in-aria-hidden.html more robust https://bugs.webkit.org/show_bug.cgi?id=221812
Reviewed by Zalan Bujtas.

It is very dependent on exact interactions between the internal testing API and the DOM.
It also seems to be partially testing a wrong thing (probably as workaround for above).

* accessibility/aria-modal-in-aria-hidden-expected.txt:
* accessibility/aria-modal-in-aria-hidden.html:

Do rendering update between every mutation and check, not just the last one.
Fix the middle test to use isIgnored too.
(isValid() is not even a function, it is a property, so this relied on getting null to pass)


Canonical link: https://commits.webkit.org/234037@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@272801 268f45cc-cd09-0410-ab3c-d52691b4dbfc