webkitpy.xcode only relevant for darwin platforms https://bugs.webkit.org/show_bug.cgi?id=221837
Reviewed by Jonathan Bedard.

The tests in webkitpy.xcode are Darwin specific and should only run on that platform.

This issue was discovered when running the tests with Python 3 on Windows which failed
a number of tests in the suite. These tests may run successfully on *nix platforms but
aren't actually relevant.

* Scripts/webkitpy/test/main.py:


Canonical link: https://commits.webkit.org/234039@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@272803 268f45cc-cd09-0410-ab3c-d52691b4dbfc