REGRESSION (r272784): [macOS Release] fast/animation/request-animation-frame-throttling-lowPowerMode.html is a flaky failure https://bugs.webkit.org/show_bug.cgi?id=221843
Reviewed by Chris Dumez.

Something about r272784 subtly changed timing, which caused there to be a little bit of a lag in RAF throttling.
I could reproduce it making the way this test measures frame rate go up to 35, but not 36.
This test is still good, though, because if you remove setLowPowerModeEnabled it spikes up to 60 or 61 frames per second.

* fast/animation/request-animation-frame-throttling-lowPowerMode-expected.txt:
* fast/animation/request-animation-frame-throttling-lowPowerMode.html:



Canonical link: https://commits.webkit.org/234060@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@272824 268f45cc-cd09-0410-ab3c-d52691b4dbfc