Crash under WorkerThreadableLoader::MainThreadBridge::notifyIsDone() https://bugs.webkit.org/show_bug.cgi?id=221906 <rdar://74225448>
Reviewed by Alex Christensen.

* loader/DocumentThreadableLoader.cpp:
(WebCore::DocumentThreadableLoader::computeIsDone):
r267227 added this function and this was the first case where the DocumentThreadableLoader
would do an async operation and ref itself during the operation. I believe this was the
source of the crash as this could cause DocumentThreadableLoader to outlive its client.
When DocumentThreadableLoader::notifyIsDone() would get called later on, m_client may be
bad. To maintain pre-r267227 behavior, we now capture a WeakPtr to |this| instead of a
Ref<>, so that we don't unnecessarily extend the lifetime of the DocumentThreadableLoader.

* loader/DocumentThreadableLoader.h:
(WebCore::DocumentThreadableLoader::clearClient):
Add new clearClient() function to DocumentThreadableLoader so that the client can clear
the raw pointer the DocumentThreadableLoader holds to it before getting destroyed. I wanted
to use WeakPtr but this did not work out due to multithreading.

* loader/WorkerThreadableLoader.cpp:
(WebCore::WorkerThreadableLoader::MainThreadBridge::destroy):
Before destroying itself, WorkerThreadableLoader::MainThreadBridge now clears the
raw pointer to it that DocumentThreadableLoader holds. This is important since
DocumentThreadableLoader is RefCounted and its lifetime may get extended past
the MainThreadBridge.

* loader/WorkerThreadableLoader.h:
Use tighter typing for clarity.


Canonical link: https://commits.webkit.org/234113@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@272880 268f45cc-cd09-0410-ab3c-d52691b4dbfc