[JSC] Enable JITCage on macOS https://bugs.webkit.org/show_bug.cgi?id=221805 <rdar://problem/74153806>
Reviewed by Mark Lam.

Source/JavaScriptCore:

We enable JITCage too on macOS if it is ARM64E.
We need to add this entitlement only when building it on macOS 120000 or higher version.
Otherwise, we cannot launch the process. This means that we need to dynamically generate entitlements file
because we must not attach this entitlement when building JSC for non 120000 macOS.

We also remove install_name for jsc binary because it broke codesigning. Previously, it was OK since we didn't
have com.apple.private.xxx, but now this broken codesigning makes JSC binary unlaunchable.

* JavaScriptCore.xcodeproj/project.pbxproj:
* Scripts/process-entitlements.sh:

Source/WebKit:

We need to add this entitlement only when building it on macOS 120000 or higher version.
Otherwise, we cannot launch the process. And we attach this entitlement only when building processes
with Apple Internal SDKs.

* Scripts/process-entitlements.sh:

Source/WTF:

Enable JIT_CAGE when macOS is 120000 or higher with ARM64E.

* wtf/PlatformEnable.h:


Canonical link: https://commits.webkit.org/234155@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@272922 268f45cc-cd09-0410-ab3c-d52691b4dbfc