RenderElement::containingBlockForAbsolutePosition may call RenderObject::containingBlock recursively https://bugs.webkit.org/show_bug.cgi?id=221976 <rdar://problem/72775667>
Reviewed by Simon Fraser.

When a RenderInline happens to be absolute positioned (this is a highly incorrect state, see webkit.org/b/221994), containingBlockForAbsolutePosition() calls containingBlock()
with |this| and in return containingBlock() calls back on containingBlockForAbsolutePosition() with the same renderer.
This patch ensures that we always call containingBlock() from containingBlockForAbsolutePosition() with an ancestor -mostly with the parent().

* rendering/RenderElement.cpp:
(WebCore::RenderElement::containingBlockForAbsolutePosition const):


Canonical link: https://commits.webkit.org/234164@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@272931 268f45cc-cd09-0410-ab3c-d52691b4dbfc