[Python-3] UnicodeDecodeError when checking for an interpreter on Windows https://bugs.webkit.org/show_bug.cgi?id=222059
Reviewed by Jonathan Bedard.

When calling open to determine which interpreter should be used Python 3 was throwing a
UnicodeDecodeError. On Python 3 an encoding of cp437 is specified since all characters are
defined in it. Python 2 does not have this problem and throws with a TypeError if an
encoding argument is passed.

* Scripts/webkitpy/common/system/executive.py:
(Executive.popen):


Canonical link: https://commits.webkit.org/234222@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@273021 268f45cc-cd09-0410-ab3c-d52691b4dbfc