Fix a few issues with RangeResponseGenerator https://bugs.webkit.org/show_bug.cgi?id=222087
Reviewed by Jer Noble.

Source/WebCore:

When it was enabled from r272908 to r273033 serveral issues were found and it was disabled until I fixed them.

The first was that RangeResponseGenerator was owned direclty by the WebCoreNSURLSession which is often deallocated on a non-main thread,
which caused all kinds of problems with the Strings and AtomStrings that are owned by its structures.
This was fixed by using WTF::DestructionThread::Main and I added assertions in the constructor and destructor.

The second was that WebCoreNSURLSessionDataTask's dealloc is casting _resource->client() to a WebCoreNSURLSessionDataTaskClient* and calling clearTask on it,
which writes null to that memory.  Before r272908 it was a good assumption that the client was always a WebCoreNSURLSessionDataTaskClient, so there was no problem.
Since then, though, it could be a MediaResourceClient.  The frustrating thing is that those two clients have the same size (32 bytes) so there was no out-of-bounds
memory writes, so ASAN wasn't seeing anything wrong.  I was just seeing 0's written randomly to my client's data.  I solved this by setting _resource to null when
the RangeResponseGenerator adopts the resource and client, so WebCoreNSURLSessionDataTask's dealloc won't do anything with it any more.

The third was that when the media stack calls WebCoreNSURLSessionDataTask's cancel, we were not removing the task from the RangeResponseGenerator's structures.
I noticed that when looking at logs, and added a call to removeTask.

Before fixing these three things, about 2/3 of the time I couldn't watch a real video on the real internet successfully.  Now, I can about 95% of the time with my
test video, which responds with an HTTP 503 about 5% of the time, and there's nothing I can do about that.  There were also some flaky tests that exercised these paths,
which have passed for me locally almost 2000 times without any failures.

* platform/network/cocoa/RangeResponseGenerator.h:
(WebCore::RangeResponseGenerator::create):
* platform/network/cocoa/RangeResponseGenerator.mm:
(WebCore::RangeResponseGenerator::RangeResponseGenerator):
(WebCore::RangeResponseGenerator::~RangeResponseGenerator):
(WebCore::RangeResponseGenerator::removeTask):
(WebCore::RangeResponseGenerator::giveResponseToTaskIfBytesInRangeReceived):
(WebCore::RangeResponseGenerator::willSynthesizeRangeResponses):
* platform/network/cocoa/WebCoreNSURLSession.h:
* platform/network/cocoa/WebCoreNSURLSession.mm:
(-[WebCoreNSURLSession initWithResourceLoader:delegate:delegateQueue:]):
(-[WebCoreNSURLSession rangeResponseGenerator]):
(-[WebCoreNSURLSessionDataTask _cancel]):
(-[WebCoreNSURLSessionDataTask resource:receivedResponse:completionHandler:]):

Tools:

* TestWebKitAPI/Tests/WebKitCocoa/MediaLoading.mm:
(TestWebKitAPI::TEST):



Canonical link: https://commits.webkit.org/234259@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@273061 268f45cc-cd09-0410-ab3c-d52691b4dbfc