Page updates and rAF callbacks should run at a frequency close to 60Hz on high refresh-rate displays https://bugs.webkit.org/show_bug.cgi?id=221673 <rdar://problem/72398605>
Reviewed by Simon Fraser.

Source/WebCore:

Up to now, page rendering updates, which can, for instance, be triggered by registering a requestAnimationFrame()
callback or having CSS Animations, CSS Transitions or Web Animations running non-accelerated, would be triggered
at the current display's refresh rate on macOS, and 60fps on iOS.

For instance, this meant that plugging in a 144Hz display on a Mac would attempt to update the page rendering at
144fps. We think it's better, for Web compatibility reasons, to try to use a page rendering update cadence that
is as close to 60fps as possible while matching a display update. This means finding the whole multiple for which
the interval between frames would be as close to that for 60Hz.

Using the 144Hz case again, this means we would shoot for 72fps, since the refresh interval for 72fps is nearest
that at 60fps than that at 48fps (another multiple of 144).

In the case where the internal preference to control whether the refresh rate for page rendering updates is
forced to 60fps, which defaults to false, we pass in the page's "nominal refresh rate" to preferredFrameInterval().
If that rate is 60fps, we use the same rate and intervals as we always have. If that rate is less than 60fps, we
use that rate as-is.

* dom/ScriptedAnimationController.cpp:
(WebCore::ScriptedAnimationController::preferredScriptedAnimationInterval const):
* page/Page.cpp:
(WebCore::Page::preferredRenderingUpdateInterval const):
* platform/graphics/AnimationFrameRate.h:
(WebCore::framesPerSecondNearestFullSpeed):
(WebCore::preferredFrameInterval):
(WebCore::preferredFramesPerSecond):

Source/WTF:

Add a new internal preference to force the refresh rate for page rendering updates to 60fps
rather than use the display's refresh rate. This setting defaults to false.

* Scripts/Preferences/WebPreferencesInternal.yaml:

Tools:

Add API tests for the methods in AnimationFrameRate.h.

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* TestWebKitAPI/Tests/WebCore/AnimationFrameRate.cpp: Added.


Canonical link: https://commits.webkit.org/234265@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@273067 268f45cc-cd09-0410-ab3c-d52691b4dbfc