[iOS] Remove access to IOSurfaceAcceleratorClient https://bugs.webkit.org/show_bug.cgi?id=222199 <rdar://problem/73518610>
Reviewed by Brent Fulgham.

Remove access to IOSurfaceAcceleratorClient in the WebContent process on iOS when the GPU process is enabled.
This patch also adds the IOKit filter from IOSurfaceRootUserClient when the GPU process is not enabled, to
the rule for IOSurfaceRootUserClient when the GPU process is enabled.

* Resources/SandboxProfiles/ios/com.apple.WebKit.WebContent.sb:


Canonical link: https://commits.webkit.org/234364@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@273169 268f45cc-cd09-0410-ab3c-d52691b4dbfc