Flaky JSC test: stress/atomic-increment-bigint64.js https://bugs.webkit.org/show_bug.cgi?id=221904
Reviewed by Keith Miller.

bytecode-cache test assumes that code can be cached once it is executed. But this
is not correct for threading tests using JSC shell's agent feature: code is running
different VM in the different thread. And we have no guarantee that this thread destroys
VM when the main thread is finished. So this is possible that bytecode used in that thread
is not committed.

We use Options::forceDiskCache's RELEASE_ASSERT only if it is in main thread.

* runtime/CodeCache.h:


Canonical link: https://commits.webkit.org/234637@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@273587 268f45cc-cd09-0410-ab3c-d52691b4dbfc